+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Prevent uploading .phps + .html files in assets manager
 - Require verification for updating user data
 - Support post field projection after content population
+- Extract width + height and colors from uploaded svg files
 
 ## 2.6.2 (2023-07-31)