+++ CHANGELOG.md
@@ -4,10 +4,11 @@
 
 - Prevent uploading .phps + .html files in assets manager
 - Require verification for updating user data
-- Support post field projection after content population
+- Support post field projection after content population (via ..{fieldname})
 - Extract width + height and colors from uploaded svg files
 - Implement HTTP caching when using response cache (via rspc parameter)
 - Resolve `{field}:locale` in filter and sort parameters
+- Optimize icon font size
 
 ## 2.6.2 (2023-07-31)