+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Resolve `{field}:locale` in filter and sort parameters
 - Optimize icon font size
 - Add missing meta fields to GraphQL model definitions @filipmiik
+- Fix long text expanding grid items
 
 ## 2.6.2 (2023-07-31)